fixes for backend
